webkit/LayoutTests/fast/canvas/canvas-methods-invalid-first-argument.html

<html>
<head>
<link rel="stylesheet" href="../js/resources/js-test-style.css">
<script src="../js/resources/js-test-pre.js"></script>
<script>
if (window.layoutTestController)
    layoutTestController.dumpAsText();

var context;
function runTest()
{
    context = document.getElementById('canvas').getContext('2d');
    shouldThrow("context.createPattern(null, 'no-repeat')", "'Error: TYPE_MISMATCH_ERR: DOM Exception 17'");
    shouldThrow("context.putImageData(null, 0, 0)", "'Error: TYPE_MISMATCH_ERR: DOM Exception 17'");
    shouldThrow("context.drawImage(null, 0, 0)", "'Error: TYPE_MISMATCH_ERR: DOM Exception 17'");
    var testContainer = document.getElementById("test-container");
    if (testContainer)
        document.body.removeChild(testContainer);
    debug('<br /><span class="pass">TEST COMPLETE</span>');
}

window.onload = runTest;
</script>
</head>
<body>
<p id="description"></p>
<div id="test-container">
    <canvas id="canvas"></canvas>
</div>
<hr/>
<div id="console"></div>
<script>
    description("This test checks that <a href=\"http://www.w3.org/TR/2dcontext/#dom-context-2d-createpattern\">createPattern</a>, " + 
                "<a href=\"http://www.w3.org/TR/2dcontext/#dom-context-2d-putimagedata\">putImageData</a>, and " +
                "<a href=\"http://www.w3.org/TR/2dcontext/#dom-context-2d-drawimage\">drawImage</a> throw a TYPE_MISMATCH_ERR " +
                "exception if their first argument is an invalid JavaScript object, as per <a href=\"http://www.w3.org/TR/2dcontext/\">" + 
                "the HTML Canvas 2D Context specification.</a>");
    var successfullyParsed = true;
</script>
</body>
</html>
